fix(database): refactor database creation and connection logic
Refactor the database module to separate the creation of the database if it does not exist
from the connection establishment process. This change improves the maintainability and
clarifies the purpose of each class method. Additionally, handle mysql.connector.Error
instead of a more general Exception to be more specific with error handling.